chore - Moved the services files into the service directory.
 - Moved database.js & eligibility-client.js for both internal and external web.
 - Adjusted all file paths to account for the moved files.
