+++ alpha/external-web/app/services/database.js
@@ -1,8 +1,8 @@
 /**
  * This file defines the database connection to Mongodb.
  */
-var logger = require('./services/bunyan-logger').logger
 var client = require('mongodb')
+var logger = require('./bunyan-logger').logger
 exports.client = client
 
 client.connect('mongodb://mongo:27017/apvs', function (error, database) {
