+++ alpha/internal-web/app/routes/claimant-details.js
@@ -1,5 +1,5 @@
 var router = require('../routes')
-var client = require('../eligibility-client')
+var client = require('../services/eligibility-client')
 var logger = require('../services/bunyan-logger').logger
 
 // Valid Statuses for a claimant application.
